refactor: 커스텀 예외 별도 클래스 분리
