+++ README.md
@@ -20,6 +20,6 @@
 ## step2 리뷰 내용
 - [x] 인스턴스 변수 불변 적용
 - [ ] 랜덤 객체 의존성 외부에서 주입받도록 변경 (전략패턴)
-- [ ] 커스텀 예외 별도 클래스로 분리
+- [x] 커스텀 예외 별도 클래스로 분리
 - [ ] domain 패키지 내부 public 함수 모든 테스트 코드 추가
 - [ ] 인스턴스 단언 람다와 의존성 주입을 통해 테스트 변경
